LPD-90977 Extract input and output variables into a shared component
