+++ modules/dxp/apps/portal-workflow/portal-workflow-kaleo-designer-web/src/main/resources/META-INF/resources/designer/js/definition-builder/diagram-builder/components/sidebar/sections/shared-components/InputOutputVariables.js
@@ -1,0 +1,79 @@
+/**
+ * SPDX-FileCopyrightText: (c) 2026 Liferay, Inc. https://liferay.com
+ * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
+ */
+
+import {ClayInput} from '@clayui/form';
+import React, {useContext, useMemo} from 'react';
+
+import {DiagramBuilderContext} from '../../../../DiagramBuilderContext';
+import {
+	formatVariablesForTextarea,
+	parseVariablesInput,
+} from '../../../../util/parseVariables';
+
+const PLACEHOLDER = '[{"name":"tone", "type":"string"}]';
+
+const InputOutputVariables = () => {
+	const {selectedItem, setSelectedItem} = useContext(DiagramBuilderContext);
+
+	const inputVariablesValue = useMemo(
+		() => formatVariablesForTextarea(selectedItem?.data?.inputVariables),
+		[selectedItem]
+	);
+
+	const outputVariablesValue = useMemo(
+		() => formatVariablesForTextarea(selectedItem?.data?.outputVariables),
+		[selectedItem]
+	);
+
+	const handleVariablesChange =
+		(field) =>
+		({target}) => {
+			if (!selectedItem) {
+				return;
+			}
+
+			setSelectedItem({
+				...selectedItem,
+				data: {
+					...selectedItem.data,
+					[field]: parseVariablesInput(target.value),
+				},
+			});
+		};
+
+	return (
+		<>
+			<label className="mt-4" htmlFor="inputVariables">
+				{Liferay.Language.get('input-variables')}
+			</label>
+
+			<ClayInput
+				className="mt-2"
+				component="textarea"
+				id="inputVariables"
+				onChange={handleVariablesChange('inputVariables')}
+				placeholder={PLACEHOLDER}
+				type="text"
+				value={inputVariablesValue}
+			/>
+
+			<label className="mt-4" htmlFor="outputVariables">
+				{Liferay.Language.get('output-variables')}
+			</label>
+
+			<ClayInput
+				className="mt-2"
+				component="textarea"
+				id="outputVariables"
+				onChange={handleVariablesChange('outputVariables')}
+				placeholder={PLACEHOLDER}
+				type="text"
+				value={outputVariablesValue}
+			/>
+		</>
+	);
+};
+
+export default InputOutputVariables;
